Authorize user api

+++ app/controllers/api/users_controller.rb
@@ -4,15 +4,19 @@ module Api
     before_action :http_authenticate_or_request
 
     def show
+      authorize(user)
       respond_with user
     end
 
     def index
-      respond_with User.page(page)
+      users = User.page(page)
+      authorize(users)
+      respond_with users
     end
 
     def create
       user = User.new(create_user_params)
+      authorize(user)
       if user.save
         render json: user
       else
@@ -21,6 +25,7 @@ module Api
     end
 
     def update
+      authorize(user)
       if user.update(update_user_params)
         render json: user
       else
@@ -29,13 +34,14 @@ module Api
     end
 
     def destroy
+      authorize(user)
       respond_with user.destroy
     end
 
     private
 
     def user
-      User.find(params[:id])
+      @user ||= User.find(params[:id])
     end
 
     def update_user_params
